requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiologger==0.7.0
aiopg==1.4.0
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
barcode==1.0.4
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
cssselect2==0.7.0
defusedxml==0.7.1
executor==23.2
fake-useragent==1.4.0
fasteners==0.19
frozenlist==1.4.1
humanfriendly==10.0
idna==3.6
imageio==2.33.1
Levenshtein==0.23.0
magic-filter==1.0.12
multidict==6.0.4
numpy==1.26.3
pillow==10.2.0
property-manager==3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pypng==0.20220715.0
python-barcode==0.15.1
python-Levenshtein==0.23.0
qrcode==7.4.2
rapidfuzz==3.6.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
tinycss2==1.2.1
typing_extensions==4.9.0
urllib3==2.1.0
verboselogs==1.7
webencodings==0.5.1
wheel==0.42.0
yarl==1.9.4
